pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillaudit"
version = "0.1.0"
description = "Security sandbox for MCP servers and AI skills"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "docker>=7.0",
    "python-dotenv>=1.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.40"]

dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
skillaudit = "skillaudit.cli:cli"
skillaudit-mcp = "skillaudit.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skillaudit*"]
